Revert "Animlogin"
