+++ src/App.css
@@ -1,21 +1,3 @@
-@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@100;300;400;500;600;700;800;900&display=swap');
-
-:root {
-  --mainbg: #23242a;
-  --babyblue: #45f3ff;
-  --blu:#45f3ff;
-  --width: 400px;
-  --height: 420px;
-}
-
-
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-
 .App {
   text-align: center;
 }
@@ -72,215 +54,4 @@
   border-bottom-left-radius: 0;
   border-top-right-radius: 2em;
   border-bottom-right-radius: 2em;
-}
-
-/* Login Form Styles */
-
-.container {
-  display: flex;
-  background-color: var(--mainbg);
-  flex-direction: column;
-  min-height: 100vh;
-  min-width: 100vw;
-  justify-content: center;
-  align-items: center;
-  font-family: 'Poppins', Arial, Helvetica, sans-serif;
-}
-
-/* BOX COMPONENT */
-
-
-.box {
-  position: relative;
-  width: var(--width);
-  height: var(--height);
-  background: #1c1c1c;
-  border-radius: 8px;
-  overflow: hidden;
-  transition: 1s;
-}
-
-@media screen and (min-width: 600px) {
-  .box {
-    scale: 1.4;
-  }
-}
-
-.box::before {
-  content: ' ';
-  position: absolute;
-  width: var(--width);
-  height: var(--height);
-  background: linear-gradient(0deg, transparent, var(--babyblue), var(--babyblue));
-  top: -50%;
-  left: -50%;
-  transform-origin: bottom right;
-  animation: animate 6s linear infinite;
-}
-
-.box::after {
-  content: ' ';
-  position: absolute;
-  width: var(--width);
-  height: var(--height);
-  background: linear-gradient(0deg, transparent, var(--babyblue), var(--babyblue));
-  top: -50%;
-  left: -50%;
-  transform-origin: bottom right;
-  animation: animate 6s linear infinite;
-  animation-delay: -3s;
-}
-
-/* FORM COMPONENT */
-
-.form {
-  position: absolute;
-  background-color: var(--mainbg);
-  inset: 2px;
-  border-radius: 8px;
-  z-index: 20;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.content {
-  width: 300px;
-  z-index: 20;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: left;
-}
-
-.form h2 {
-  color: #8f8f8f;
-  font-size: 2em;
-  font-weight: 700;
-}
-
-.inputBox {
-  width: 300px; 
-  position: relative;
-  transition: 100ms;
-  margin-top: 20px;
-}
-
-input:-webkit-autofill,
-.inputBox input {
-  position: relative;
-  width: 100%;
-  padding: 20px 10px 10px;
-  background-color: transparent;
-  border: none;
-  outline: none;
-  color: #23242a;
-  z-index: 10;
-  margin: .8em 0 0 0;
-  /* font-size: 1.1em; */
-  font-weight: 300;
-}
-
-.inputBox span {
-  position: absolute;
-  left: 0;
-  padding: 20px 0px 10px;
-  font-size: 1em;
-  color: #8f8f8f;
-  font-weight: 300;
-  pointer-events: none;
-  letter-spacing: 0.05em;
-  transition: 0.5s ease;
-}
-
-
-.inputBox i {
-  position: absolute;
-  left: 0;
-  bottom: 0;
-  width: 100%;
-  height: 1px;
-  border-radius: 8px;
-  transition: 100ms;
-  background-color: var(--babyblue);
-}
-
-
-.inputBox input:valid ~ span,
-.inputBox input:focus ~ span {
-  color: var(--babyblue);
-  transform: translateY(-24px);
-}
-
-.inputBox input:valid ~ i,
-.inputBox input:focus ~ i {
-  height: 2.5em;
-}
-
-/* LINKS */
-
-.links {
-  display: flex;
-  width: 100%;
-  justify-content: flex-end;
-  padding: 1em 0;
-  z-index: 10;
-}
-
-.links a,
-.links button,
-.links p {
-  text-decoration: none;
-  font-weight: 300;
-  font-size: 0.75em;
-  color: #8f8f8f;
-  z-index: 10;
-}
-
-.links p:hover {
-  color: var(--babyblue);
-  cursor: pointer;
-  transition: 1s ease;
-}
-
-.links a:hover,
-.links a:nth-child(2) {
-  color: var(--babyblue);
-}
-
-.links a:nth-child(2):hover {
-  font-weight: bold;
-}
-
-.submit {
-  width: 80%;
-  align-self: center;
-  padding: 1em;
-  font-weight: bold;
-  font-family: inherit;
-  background-color: var(--babyblue);
-  border: none;
-  border-radius: .4em;
-  z-index: 20;
-  margin-top: 1em;
-}
-
-.submit:hover {
-  cursor: pointer;
-}
-
-/* ANIMATIONS */
-
-@keyframes animate {
-  0% {
-    rotate: 0deg;
-    -webkit-transform: rotate(0deg);
-    -moz-transform: rotate((0deg));
-  }
-  100% {
-    rotate: 360deg;
-    -webkit-transform: rotate(360deg);
-    -moz-transform: rotate(360deg);
-  }
 }
